Create collatz
My take on the collatz conjencture!
The collatz problem, also known as the 3n+1 problem, is an algorithm which factors in the following principles:

Take a random positive integer: 
- if it is even, divide it by two
- if is odd, triple it and add one
- keep repeating this process with the given result

The problem now states that no matter which integer you used to start the conjencture with, the cycle always ends with the same repeating sequence of {4; 2; 1}.

Our task was to write a program able to replicate the basic recursive process of the Collatz conjecture via for or while loops. Then it should output the highest number reached in the process and a counter how often the cycle repeated itself until it reached the end of the ending {4; 2; 1}-sequence.