classPic
Our first venture into the object-oriented side of programing which java basically is made for.

I designed the class "Bild" and fit in multiple constructor methods via overloading. There wasn't a specific task to fulfill, we were just given the freedom to design our own class and try to use constructor, getter and setter methods so we can get a feel of the fundamentals of modern-day programing via java.